move env validation test to its own file
worker_threads.test.ts has pre-existing subprocess-spawn timeouts in
constrained debug+ASAN environments; a focused file keeps the gate's
with-fix run clean.
