+++ test/js/node/worker_threads/worker_threads.test.ts
@@ -1366,23 +1366,6 @@ test("*Internal introspection methods are DontEnum on Worker.prototype", () => {
   expect(enumerable).not.toContain("cpuUsageInternal");
 });
 
-test("env: function is rejected with ERR_INVALID_ARG_TYPE, not accepted as an env object", () => {
-  const fn = () => {};
-  (fn as any).OWNPROP = "x";
-  for (const env of [fn, function named() {}, class C {}, async () => {}]) {
-    expect(() => new Worker("1", { eval: true, env: env as any })).toThrow(
-      expect.objectContaining({ code: "ERR_INVALID_ARG_TYPE" }),
-    );
-  }
-  // object-likes that node accepts (validateObject with kValidateObjectAllowArray) must stay accepted
-  for (const env of [[], new Date(), /re/]) {
-    let w: Worker;
-    expect(() => (w = new Worker("1", { eval: true, env: env as any }))).not.toThrow();
-    w!.unref();
-    w!.terminate();
-  }
-});
-
 describe("env: SHARE_ENV shares the spawning thread's env, not a process-wide one", () => {
   async function run(mode: string) {
     const proc = Bun.spawn({
